Correção da branch e ajustes nos endpoints de produtos
