--- src/main/java/com/sedocefosse/backend/controller/AdminProductController.java
@@ -1,28 +1,0 @@
-package com.sedocefosse.backend.controller;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import com.sedocefosse.backend.model.Product;
-import com.sedocefosse.backend.service.ProductService;
-
-@RestController 
-@RequestMapping("/admin/products") 
-public class AdminProductController {
-
-    private final ProductService productService;
-
-    public AdminProductController(ProductService productService) {
-        this.productService = productService;
-    }
-
-    @PostMapping
-    public ResponseEntity<Product> createProduct(@RequestBody Product product) {        
-        Product createdProduct = productService.create(product);
-        return new ResponseEntity<>(createdProduct, HttpStatus.CREATED);
-    }
-}
