--- src/main/java/com/sedocefosse/backend/model/Product.java
@@ -1,23 +1,0 @@
-package com.sedocefosse.backend.model;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import lombok.Data;
-
-@Data 
-@Entity 
-@Table(name = "products") 
-public class Product {
-
-    @Id 
-    @GeneratedValue(strategy = GenerationType.IDENTITY) 
-    private Long id;
-    private String name;
-    private String description;
-    private double price;
-    private String imageUrl;
-    private String restricaoAlimentar;
-}
