--- src/main/java/com/sedocefosse/backend/repository/ProductRepository.java
@@ -1,11 +1,0 @@
-package com.sedocefosse.backend.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import com.sedocefosse.backend.model.Product;
-
-@Repository 
-public interface ProductRepository extends JpaRepository<Product, Long> {
-    
-}
