--- src/main/java/com/sedocefosse/backend/service/ProductService.java
@@ -1,9 +1,0 @@
-package com.sedocefosse.backend.service;
-
-import com.sedocefosse.backend.model.Product;
-
-public interface ProductService {
-    
-    Product create(Product product);
-
-}
