--- src/main/java/com/sedocefosse/backend/service/ProductServiceImpl.java
@@ -1,32 +1,0 @@
-package com.sedocefosse.backend.service;
-
-
-import org.springframework.stereotype.Service;
-import com.sedocefosse.backend.model.Product;
-import com.sedocefosse.backend.repository.ProductRepository;
-
-@Service 
-public class ProductServiceImpl implements ProductService {
-
-    private final ProductRepository productRepository;
-
-    public ProductServiceImpl(ProductRepository productRepository) {
-        this.productRepository = productRepository;
-    }
-
-    @Override
-    public Product create(Product product) {
-        return productRepository.save(product);
-    }
-}
-
-//Teste manual a ser feito no Postman
-//POST http://localhost:8080/api/products
-//Body (JSON):
-// {
-//     "name": "Produto Exemplo",
-//     "description": "Descrição do Produto Exemplo",
-//     "price": 19.99,
-//     "imageUrl": "http://exemplo.com/imagem.jpg",
-//     "restricaoAlimentar": "Sem glúten"
-
